Completely redesigned
Now a single C program that interfaces with the Data Hub as a
non-administrative app.  Connections from this app to the LEDs
are left to some other administrative app to do.

Slowed down the default continuous mode and changed the pattern
to just cycle through the different software-controlled LEDs.

Made it possible to disable the continuous mode and configure
the continuous mode cycle speed through the Data Hub.

Added a triggered mode, with a trigger and configuration settings
available in the Data Hub.